# Implement TEN VAD pre-check system with advanced post-processing