--- a/.claude/commands/execute-select.md
+++ b/.claude/commands/execute-select.md
@@ -1,0 +1,3 @@
+List files in (ordered by date, descending) in `documentation/implementation-plans` and list for selection (eg. `State Management Update (2025-01-28 12:00)`)
+
+Then run the `.claude/execute.md` command after the user selects using that selected plan as the argument