--- a/.claude/commands/execute.md
+++ b/.claude/commands/execute.md
@@ -1,13 +1,11 @@
 PLAN: $ARGUMENTS
 
 If PLAN is file reference
-- Think hard and implement $ARGUMENTS
+- Implement $ARGUMENTS
 
 If PLAN is "l"
-- Think hard and implement plan with the latest timestamp in `documentation/implementation-plans`
+- Implement plan with the latest timestamp in `documentation/implementation-plans`
 
-If PLAN is empty
-- List files in (ordered by date, descending) in `documentation/implementation-plans` and list for selection (eg. `State Management Update (2025-01-28 12:00)`)
-- Then think hard and implement user's selection
+Think hard
 
 @documentation/plan-execution.md