--- a/.claude/commands/write-plan.md
+++ b/.claude/commands/write-plan.md
@@ -1,5 +1,5 @@
 Create implementation plan for: $ARGUMENTS
 
-Think hard
+Think harder
 
 @documentation/plan-writing.md