--- a/documentation/plan-writing.md
+++ b/documentation/plan-writing.md
@@ -19,9 +19,6 @@ Each implementation plan should follow this structure:
 # [Feature Name] Implementation Plan
 
 As a *[User]* I want **[Feature]** so [Reason]
-
-**Note: Execute one line-item at a time and confirm with user before proceeding to next**
-**Note: Update this document as you go, checking off tasks and taking notes needed for later**
 ```
 
 ### Current State Analysis