--- a/requirements.txt
+++ b/requirements.txt
@@ -16,3 +16,6 @@ Pillow>=10.0.0
 
 # Cache management dependencies
 hf-xet>=1.1.5
+
+# Voice Activity Detection (VAD)
+git+https://github.com/TEN-framework/ten-vad.git