… documentation/implementation-plans/2025-08-05-1232-vad-precheck-short-recordings.md
… documentation/implementation-plans/2025-08-05-1236-vad-trailing-silence-detection.md
… documentation/implementation-plans/2025-08-05-1547-vad-postprocessing-pipeline-fix.md
… documentation/research/VAD-libraries.md
… refactor.md