## Summary

Implements a comprehensive Voice Activity Detection (VAD) pre-check system using TEN VAD to prevent Whisper hallucinations on silent recordings and improve performance.

• **TEN VAD Integration**: High-precision speech detection with 5-57ms processing time
• **Advanced Post-Processing**: Hysteresis thresholds and consecutive frame filtering eliminate false positives  
• **Performance Optimization**: Silent recordings skip transcription (~1200ms → ~50ms processing time)
• **Clean User Experience**: Removed verbose debugging output, improved console messaging
• **Robust Detection**: Prevents flickering on borderline audio, filters brief noise/music

## Key Features

### TEN VAD Pre-Check System
- Processes audio in 256-sample chunks (16ms frames) at 16kHz
- Uses hysteresis thresholds (onset=0.7, offset=0.55) to prevent rapid on/off switching
- Implements 100ms minimum duration filtering to ignore brief noise spikes
- Returns binary "any speech detected?" decision for holistic recording analysis

### Post-Processing Pipeline  
- **Probability Buffering**: Collects all frame probabilities before making decisions
- **Hysteresis Logic**: Stable speech state transitions prevent flickering on borderline audio
- **Consecutive Frame Filtering**: Requires sustained speech segments (100ms minimum)
- **Standards Compliance**: Follows WebRTC/Silero/PyAnnote VAD approaches

### Performance Improvements
- Silent recordings: ~1200ms → ~50ms processing time (96% reduction)
- Short recordings with speech: No performance impact
- Prevents Whisper hallucinations ("Thank you for watching", etc.) on silence
- Maintains full transcription accuracy for legitimate speech

### User Experience Enhancements
- Clean console output with consistent indentation and emoji indicators
- Removed verbose VAD debugging (maintained in logs for troubleshooting)
- Informative messages: "No speech detected, skipping transcription"
- Professional visual hierarchy with proper status indicators

## Technical Implementation

### Files Modified
- `src/whisper_engine.py`: TEN VAD integration and post-processing pipeline
- `src/audio_recorder.py`: Hardcoded 16kHz sample rate (optimal for TEN VAD)
- `src/config_manager.py`: VAD configuration validation
- `src/state_manager.py`: Console message improvements
- `src/clipboard_manager.py`: Cleaned up paste operation logging
- `config.yaml`: Added VAD pre-check configuration options

### Architecture
```
Audio Input (16kHz) → TEN VAD (256-sample chunks) → Post-Processing Pipeline → Binary Decision
                                                     ↓
                              Hysteresis + Consecutive Frames + Duration Filtering
                                                     ↓
                               Speech Detected: Continue to Whisper Transcription
                               Silence Detected: Skip transcription, return immediately
```

## Test Plan

- [x] Silent recordings correctly detected and skipped
- [x] Speech recordings properly transcribed without regression
- [x] Background music/noise filtered out (tested with quiet background music)
- [x] Brief speech segments detected reliably
- [x] Console output clean and informative
- [x] Performance improvements verified (96% reduction for silent recordings)

🤖 Generated with [Claude Code](https://claude.ai/code)